+++ tdmclient/tools/gui.py
@@ -500,6 +500,7 @@ Options:
   --debug n      display diagnostic information (0=none, 1=basic, 2=more, 3=verbose)
   --help         display this help message and exit
   --language=L   programming language (aseba or python); default=automatic
+  --password=PWD specify password for remote tdm
   --robotid=I    robot id; default=any
   --robotname=N  robot name; default=any
   --tdmaddr=H    tdm address (default: localhost or from zeroconf)
@@ -513,6 +514,7 @@ if __name__ == "__main__":
     language = None  # auto
     tdm_addr = None
     tdm_port = None
+    password = None
     robot_id = None
     robot_name = None
 
@@ -523,6 +525,7 @@ if __name__ == "__main__":
                                               "debug=",
                                               "help",
                                               "language=",
+                                              "password=",
                                               "robotid=",
                                               "robotname=",
                                               "tdmaddr=",
@@ -539,6 +542,8 @@ if __name__ == "__main__":
             debug = int(val)
         elif arg == "--language":
             language = val
+        elif arg == "--password":
+            password = val
         elif arg == "--robotid":
             robot_id = val
         elif arg == "--robotname":
@@ -549,6 +554,7 @@ if __name__ == "__main__":
             tdm_port = int(val)
 
     win = VariableTableWindow(tdm_addr=tdm_addr, tdm_port=tdm_port,
+                              password=password,
                               node_id=robot_id, node_name=robot_name,
                               language=language, debug=debug)
     win.connect()