+++ tdmclient/tools/watch.py
@@ -17,6 +17,7 @@ Watch information on robot sent by tdm
 Options:
   --debug=n      display diagnostic info (0=none, 1=basic, 2=more, 3=verbose)
   --help         display this help message and exit
+  --password=PWD specify password for remote tdm
   --robotid=I    robot id; default=any
   --robotname=N  robot name; default=any
   --tdmaddr=H    tdm address (default: localhost or from zeroconf)
@@ -29,6 +30,7 @@ if __name__ == "__main__":
     debug = 2  # display all messages received from the tdm by default
     tdm_addr = None
     tdm_port = None
+    password = None
     robot_id = None
     robot_name = None
 
@@ -38,6 +40,7 @@ if __name__ == "__main__":
                                           [
                                               "debug=",
                                               "help",
+                                              "password=",
                                               "robotid=",
                                               "robotname=",
                                               "tdmaddr=",
@@ -52,6 +55,8 @@ if __name__ == "__main__":
             sys.exit(0)
         elif arg == "--debug":
             debug = int(val)
+        elif arg == "--password":
+            password = val
         elif arg == "--robotid":
             robot_id = val
         elif arg == "--robotname":
@@ -65,7 +70,9 @@ if __name__ == "__main__":
         help(file=sys.stderr)
         sys.exit(1)
 
-    with ClientAsync(tdm_addr=tdm_addr, tdm_port=tdm_port, debug=debug) as client:
+    with ClientAsync(tdm_addr=tdm_addr, tdm_port=tdm_port,
+                     password=password,
+                     debug=debug) as client:
 
         async def prog():
             node = await client.wait_for_node(node_id=robot_id, node_name=robot_name)